requirements.txt
# Core Django
Django==4.2.0

# REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
PyMySQL==1.1.0  # 순수 Python MySQL 드라이버 (설치 간단)
cryptography==41.0.5  # MySQL 8.0 인증에 필요

# Image Processing
Pillow==10.0.0

# Celery & Redis
celery==5.3.0
redis==5.0.0
django-redis==5.4.0
django-celery-beat==2.5.0

# CORS
django-cors-headers==4.3.0

# Environment Variables
python-decouple==3.8

# Excel Processing (for data import)
openpyxl==3.1.2

# Development Tools (개발 시 필요)
django-debug-toolbar==4.2.0
ipython==8.14.0

# Testing
factory-boy==3.3.0
faker==19.3.0
freezegun==1.2.2

# Code Quality
flake8==6.1.0
black==23.7.0
isort==5.12.0

# Production Server
gunicorn==21.2.0